Web Update

+++ WEB/main/views.py
@@ -1,10 +1,11 @@
 from django.shortcuts import redirect, render
+from django.contrib import messages
 # View에 Model(Post 게시글) 가져오기
 from .models import Post
 
 # Create your views here.
-def index(request):
-    return render(request,'main/index.html')
+def credit(request):
+    return render(request,'main/credit.html')
 
 # index.html 페이지를 부르는 index 함수
 def index(request):
@@ -25,24 +26,18 @@ def posting(request, pk):
 
 def new_post(request):
     if request.method == 'POST':
-        if request.POST['mainphoto']:
-            new_article=Post.objects.create(
-                postname=request.POST['postname'],
-                contents=request.POST['contents'],
-                mainphoto=request.POST['mainphoto'],
-            )
-        else:
-            new_article=Post.objects.create(
-                postname=request.POST['postname'],
-                contents=request.POST['contents'],
-                mainphoto=request.POST['mainphoto'],
-            )
+        new_article=Post.objects.create(
+            postname=request.POST['postname'],
+            contents=request.POST['contents'],
+            name=request.POST['name'],
+            password=request.POST['password'],
+        )
         return redirect('/blog/')
     return render(request, 'main/new_post.html')
 
 def remove_post(request, pk):
     post = Post.objects.get(pk=pk)
-    if request.method == 'POST':
+    if request.method == 'POST' and request.POST['password'] == post.password:
         post.delete()
         return redirect('/blog/')
     return render(request, 'main/remove_post.html', {'Post': post})
